VL/pytorch-yolo/requirements.txt
absl-py == 0.7.1
alabaster == 0.7.12
asn1crypto == 0.24.0
astor == 0.7.1
astroid == 2.2.5
babel == 2.7.0
backcall == 0.1.0
bleach == 3.1.0
cloudpickle == 1.2.1
cryptography == 2.7
cython == 0.29.13
decorator == 4.4.0
docutils == 0.15.2
entrypoints == 0.3
gast == 0.2.2
imageio ==2.5.0
imageio-ffmpeg == 0.3.0
ipython == 7.7.0
ipython_genutils == 0.2.0
jsonschema == 3.0.2
markdown == 3.1.1
numpy == 1.17.0
olefile == 0.46
pandas == 0.25.0
pandoc == 1.0.2
parso == 0.5.1
path.py == 12.0.1
pickleshare == 0.7.5
prompt_toolkit == 2.0.9
protobuf3 == 0.2.1
psutil == 5.6.3
pyOpenSSL == 19.0.0
scipy == 1.3.1
spyder == 3.3.6
tensorboard == 1.14.0
tensorflow == 1.14.0
tensorflow-gpu == 1.14.0
wheel == 0.33.4
wrapt == 1.11.2
ffmpeg == 1.1.0
keras == 2.2.4
keras-applications == 1.0.8
keras-preprocessing == 1.1.0
kivy == 1.11.1
kiwisolver == 1.1.0
mako == 1.1.0
pillow == 6.1.0
pyyaml == 5.1.2
theano == 1.0.4
matplotlib == 3.1.1
torch == 1.2.0
json262 == 0.2.0
qutils == 0.5.14
